auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["SkMishra <sanatkumarmishra02@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
motor = "^3.6.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.5.2"
pydantic = {extras = ["email"], version = "^2.9.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
